requirements.txt
streamlit>=1.31.0
google-cloud-bigquery>=3.11.0
google-cloud-discoveryengine>=0.11.0
google-genai>=0.2.0
google-auth>=2.23.0
